tests and small chages
